improve styles and process
