+++ components/home-page.tsx
@@ -1,17 +1,10 @@
 "use client";
 
 import { useState } from "react";
-import Link from "next/link";
+
 import { useRouter } from "next/navigation";
 import { Button } from "@/components/ui/button";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
+
 import { Input } from "@/components/ui/input";
 import {
   Dialog,
@@ -22,21 +15,15 @@ import {
   DialogTitle,
   DialogTrigger,
 } from "@/components/ui/dialog";
-import {
-  FileIcon,
-  PlusIcon,
-  PenToolIcon,
-  Loader2,
-  Search,
-  Clock,
-} from "lucide-react";
+import { PlusIcon, PenToolIcon, Loader2, Search } from "lucide-react";
 import { trpc } from "@/lib/trpc/client";
-import { LoadingSpinner } from "@/components/loading-spinner";
+
 import { toast } from "sonner";
-import { DateTime } from "luxon";
-import { ScrollArea } from "@/components/ui/scroll-area";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import { DocumentClient, Error } from "@/lib/document-store";
+
+import { DocumentClient } from "@/lib/document-store";
+
+import { Card } from "./ui/card";
+import { TabSections } from "./homepage/tab-sections";
 
 interface HomePageProps {
   initialDocuments: DocumentClient[];
@@ -51,7 +38,7 @@ export default function HomePage({ initialDocuments }: HomePageProps) {
   const {
     data: documents,
     isLoading,
-    error,
+
     refetch,
   } = trpc.document.getDocuments.useQuery(undefined, {
     refetchOnWindowFocus: false,
@@ -84,7 +71,9 @@ export default function HomePage({ initialDocuments }: HomePageProps) {
   );
 
   const recentDocuments = [...(filteredDocuments || [])].sort(
-    (a, b) => new Date(b.updatedAt).getTime() - new Date(a.updatedAt).getTime()
+    (a, b) =>
+      new Date(b.updatedAt || "").getTime() -
+      new Date(a.updatedAt || "").getTime()
   );
 
   return (
@@ -149,8 +138,6 @@ export default function HomePage({ initialDocuments }: HomePageProps) {
             filteredDocuments={filteredDocuments}
             recentDocuments={recentDocuments}
             isLoading={isLoading}
-            error={error}
-            refetch={refetch}
             searchQuery={searchQuery}
           />
         </Card>
@@ -207,220 +194,5 @@ const CreateDocumentDialog = ({
         </DialogFooter>
       </DialogContent>
     </Dialog>
-  );
-};
-
-export interface TabContentProps {
-  documents: DocumentClient[];
-  filteredDocuments: DocumentClient[];
-  recentDocuments: DocumentClient[];
-  isLoading: boolean;
-  error: Error;
-  refetch: () => void;
-  searchQuery: string;
-}
-
-const getTabContent = (
-  value: string,
-  props: TabContentProps
-): React.ReactNode => {
-  const contents: Record<string, React.ReactNode> = {
-    all: (
-      <DocumentList
-        documents={props.filteredDocuments}
-        isLoading={props.isLoading}
-        error={props.error}
-        refetch={props.refetch}
-        searchQuery={props.searchQuery}
-      />
-    ),
-    recent: (
-      <RecentDocuments
-        documents={props.recentDocuments.slice(0, 6)}
-        isLoading={props.isLoading}
-      />
-    ),
-  };
-
-  return contents[value] ?? null;
-};
-
-interface TabSectionsProps extends TabContentProps {
-  defaultTab?: string;
-}
-
-export function TabSections({
-  defaultTab = "all",
-  ...props
-}: TabSectionsProps) {
-  return (
-    <Tabs defaultValue={defaultTab} className="flex-grow flex flex-col">
-      <div className="px-6 pt-4 border-b">
-        <TabsList>
-          <TabsTrigger value="all">All Documents</TabsTrigger>
-          <TabsTrigger value="recent">Recent</TabsTrigger>
-        </TabsList>
-      </div>
-
-      <div className="flex-grow flex flex-col overflow-hidden">
-        {["all", "recent"].map((tabValue) => (
-          <TabsContent
-            key={tabValue}
-            value={tabValue}
-            className="flex-grow flex flex-col p-0 m-0 h-full"
-          >
-            <ScrollArea className="h-[calc(100vh-320px)] min-h-[400px] dark:bg-zinc-900">
-              <div className="p-6">{getTabContent(tabValue, props)}</div>
-            </ScrollArea>
-          </TabsContent>
-        ))}
-      </div>
-    </Tabs>
-  );
-}
-
-const DocumentList = ({
-  documents,
-  isLoading,
-  error,
-  refetch,
-  searchQuery,
-}: {
-  documents: DocumentClient[];
-  isLoading: boolean;
-  error: Error;
-  refetch: () => void;
-  searchQuery: string;
-}) => {
-  return (
-    <TabsContent value="all" className="flex-grow flex flex-col p-0 m-0 h-full">
-      <ScrollArea className="h-[calc(100vh-320px)] min-h-[400px] dark:bg-zinc-900">
-        <div className="p-6">
-          {isLoading ? (
-            <div className="flex justify-center py-12">
-              <LoadingSpinner size="lg" text="Loading your documents..." />
-            </div>
-          ) : error ? (
-            <div className="text-center py-12">
-              <p className="text-muted-foreground mb-4">
-                Failed to load documents
-              </p>
-              <Button onClick={() => refetch()} variant="outline">
-                Try Again
-              </Button>
-            </div>
-          ) : documents && documents.length > 0 ? (
-            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
-              {documents.map((doc) => (
-                <Link
-                  href={`/editor/${doc.id}`}
-                  key={doc.id}
-                  className="block group"
-                >
-                  <Card className="h-full transition-all duration-300 hover:shadow-lg dark:hover:shadow-primary/10 hover:-translate-y-1 border-border/50 overflow-hidden">
-                    <CardHeader className="pb-2">
-                      <CardTitle className="group-hover:text-primary transition-colors line-clamp-1">
-                        {doc.title}
-                      </CardTitle>
-                      <CardDescription className="flex items-center gap-1">
-                        <Clock className="h-3 w-3" />
-                        {DateTime.fromISO(doc.updatedAt).toRelative()}
-                      </CardDescription>
-                    </CardHeader>
-                    <CardContent className="flex items-center justify-center py-8">
-                      <div className="relative group-hover:scale-110 transition-transform duration-300">
-                        <div className="absolute inset-0 bg-primary/10 dark:bg-primary/5 rounded-full blur-xl"></div>
-                        <FileIcon className="h-16 w-16 text-primary relative z-10" />
-                      </div>
-                    </CardContent>
-                    <CardFooter className="text-sm text-muted-foreground bg-muted/30 dark:bg-muted/10 py-2">
-                      Click to open in editor
-                    </CardFooter>
-                  </Card>
-                </Link>
-              ))}
-            </div>
-          ) : (
-            <div className="text-center py-12 border rounded-lg bg-muted/5 dark:bg-muted/10">
-              <FileIcon className="h-12 w-12 text-muted-foreground/50 mx-auto mb-4" />
-              <h3 className="text-lg font-medium mb-2">
-                {searchQuery
-                  ? "No matching documents found"
-                  : "No documents yet"}
-              </h3>
-              <p className="text-muted-foreground mb-4">
-                {searchQuery
-                  ? "Try a different search term or create a new document"
-                  : "Create your first document to get started"}
-              </p>
-            </div>
-          )}
-        </div>
-      </ScrollArea>
-    </TabsContent>
-  );
-};
-
-const RecentDocuments = ({
-  documents,
-  isLoading,
-}: {
-  documents: DocumentClient[];
-  isLoading: boolean;
-}) => {
-  return (
-    <TabsContent
-      value="recent"
-      className="flex-grow flex flex-col p-0 m-0 h-full"
-    >
-      <ScrollArea className="h-[calc(100vh-320px)] min-h-[400px] dark:bg-zinc-900">
-        <div className="p-6">
-          {isLoading ? (
-            <div className="flex justify-center py-12">
-              <LoadingSpinner size="lg" text="Loading your documents..." />
-            </div>
-          ) : documents && documents.length > 0 ? (
-            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
-              {documents.slice(0, 6).map((doc) => (
-                <Link
-                  href={`/editor/${doc.id}`}
-                  key={doc.id}
-                  className="block group"
-                >
-                  <Card className="h-full transition-all duration-300 hover:shadow-lg dark:hover:shadow-primary/10 hover:-translate-y-1 border-border/50 overflow-hidden">
-                    <CardHeader className="pb-2">
-                      <CardTitle className="group-hover:text-primary transition-colors line-clamp-1">
-                        {doc.title}
-                      </CardTitle>
-                      <CardDescription className="flex items-center gap-1">
-                        <Clock className="h-3 w-3" />
-                        {DateTime.fromISO(doc.updatedAt).toRelative()}
-                      </CardDescription>
-                    </CardHeader>
-                    <CardContent className="flex items-center justify-center py-8">
-                      <div className="relative group-hover:scale-110 transition-transform duration-300">
-                        <div className="absolute inset-0 bg-primary/10 dark:bg-primary/5 rounded-full blur-xl"></div>
-                        <FileIcon className="h-16 w-16 text-primary relative z-10" />
-                      </div>
-                    </CardContent>
-                    <CardFooter className="text-sm text-muted-foreground bg-muted/30 dark:bg-muted/10 py-2">
-                      Click to open in editor
-                    </CardFooter>
-                  </Card>
-                </Link>
-              ))}
-            </div>
-          ) : (
-            <div className="text-center py-12 border rounded-lg bg-muted/5 dark:bg-muted/10">
-              <Clock className="h-12 w-12 text-muted-foreground/50 mx-auto mb-4" />
-              <h3 className="text-lg font-medium mb-2">No recent documents</h3>
-              <p className="text-muted-foreground mb-4">
-                Your recently edited documents will appear here
-              </p>
-            </div>
-          )}
-        </div>
-      </ScrollArea>
-    </TabsContent>
   );
 };
