--- components/homepage/gallery/document-dialog.tsx
@@ -1,64 +1,0 @@
-import { Button } from "@/components/ui/button";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
-import { Input } from "@/components/ui/input";
-import { Loader2 } from "lucide-react";
-
-interface CreateDocumentDialogProps {
-  isOpen: boolean;
-  setIsOpen: (open: boolean) => void;
-  title: string;
-  setTitle: (title: string) => void;
-  onSubmit: () => void;
-  isPending: boolean;
-}
-
-export function CreateDocumentDialog({
-  isOpen,
-  setIsOpen,
-  title,
-  setTitle,
-  onSubmit,
-  isPending,
-}: CreateDocumentDialogProps) {
-  return (
-    <Dialog open={isOpen} onOpenChange={setIsOpen}>
-      <DialogContent className="sm:max-w-[425px]">
-        <DialogHeader>
-          <DialogTitle>Create New Document</DialogTitle>
-          <DialogDescription>
-            Give your document a title to get started.
-          </DialogDescription>
-        </DialogHeader>
-        <Input
-          value={title}
-          onChange={(e) => setTitle(e.target.value)}
-          placeholder="Document title"
-          className="my-4"
-        />
-        <DialogFooter>
-          <Button
-            onClick={onSubmit}
-            disabled={isPending}
-            className="w-full sm:w-auto"
-          >
-            {isPending ? (
-              <>
-                <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-                Creating...
-              </>
-            ) : (
-              "Create Document"
-            )}
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  );
-}
